Revert "Rename `api.components.selections` to `api.components.selects`"
This reverts commit 5564447075a48b9359fb2106d8f62e96794f1792.
