+++ src/main/java/net/dv8tion/jda/api/components/actionrow/ActionRowChildComponent.java
@@ -18,8 +18,8 @@ package net.dv8tion.jda.api.components.actionrow;
 
 import net.dv8tion.jda.api.components.Component;
 import net.dv8tion.jda.api.components.buttons.Button;
-import net.dv8tion.jda.api.components.selects.EntitySelectMenu;
-import net.dv8tion.jda.api.components.selects.StringSelectMenu;
+import net.dv8tion.jda.api.components.selections.EntitySelectMenu;
+import net.dv8tion.jda.api.components.selections.StringSelectMenu;
 
 /**
  * Represents a component that can be added to an {@link ActionRow}, this includes:
