+++ src/main/java/net/dv8tion/jda/api/components/selections/SelectMenu.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.dv8tion.jda.api.components.selects;
+package net.dv8tion.jda.api.components.selections;
 
 import net.dv8tion.jda.annotations.ForRemoval;
 import net.dv8tion.jda.annotations.ReplaceWith;
@@ -38,7 +38,7 @@ import java.util.Collection;
  * Other users cannot see the choices selected, and they will disappear when the client restarts or the message is reloaded.
  *
  * <p>This is a generic interface for all types of select menus.
- * <br>You can use {@link EntitySelectMenu#create(String, Collection)} to create a select menu of Discord entities such as {@link net.dv8tion.jda.api.components.selects.EntitySelectMenu.SelectTarget#USER users}.
+ * <br>You can use {@link EntitySelectMenu#create(String, Collection)} to create a select menu of Discord entities such as {@link net.dv8tion.jda.api.components.selections.EntitySelectMenu.SelectTarget#USER users}.
  * <br>Alternatively, you can use {@link StringSelectMenu#create(String)} to create a select menu of up to {@value #OPTIONS_MAX_AMOUNT} pre-defined strings to pick from.
  *
  * @see StringSelectMenu
