+++ src/main/java/net/dv8tion/jda/api/events/interaction/component/EntitySelectInteractionEvent.java
@@ -17,7 +17,7 @@
 package net.dv8tion.jda.api.events.interaction.component;
 
 import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.components.selects.EntitySelectMenu;
+import net.dv8tion.jda.api.components.selections.EntitySelectMenu;
 import net.dv8tion.jda.api.entities.IMentionable;
 import net.dv8tion.jda.api.entities.Mentions;
 import net.dv8tion.jda.api.interactions.components.selections.EntitySelectInteraction;
