+++ src/main/java/net/dv8tion/jda/api/events/interaction/component/GenericSelectMenuInteractionEvent.java
@@ -17,7 +17,7 @@
 package net.dv8tion.jda.api.events.interaction.component;
 
 import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.components.selects.SelectMenu;
+import net.dv8tion.jda.api.components.selections.SelectMenu;
 import net.dv8tion.jda.api.interactions.components.selections.SelectMenuInteraction;
 
 import javax.annotation.Nonnull;
