+++ src/main/java/net/dv8tion/jda/api/events/interaction/component/StringSelectInteractionEvent.java
@@ -17,7 +17,7 @@
 package net.dv8tion.jda.api.events.interaction.component;
 
 import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.components.selects.StringSelectMenu;
+import net.dv8tion.jda.api.components.selections.StringSelectMenu;
 import net.dv8tion.jda.api.interactions.components.selections.StringSelectInteraction;
 
 import javax.annotation.Nonnull;
