+++ src/main/java/net/dv8tion/jda/internal/interactions/components/selections/EntitySelectInteractionImpl.java
@@ -16,7 +16,7 @@
 
 package net.dv8tion.jda.internal.interactions.components.selections;
 
-import net.dv8tion.jda.api.components.selects.EntitySelectMenu;
+import net.dv8tion.jda.api.components.selections.EntitySelectMenu;
 import net.dv8tion.jda.api.entities.IMentionable;
 import net.dv8tion.jda.api.entities.Mentions;
 import net.dv8tion.jda.api.interactions.components.selections.EntitySelectInteraction;
