+++ src/main/java/net/dv8tion/jda/internal/interactions/components/selections/StringSelectInteractionImpl.java
@@ -16,7 +16,7 @@
 
 package net.dv8tion.jda.internal.interactions.components.selections;
 
-import net.dv8tion.jda.api.components.selects.StringSelectMenu;
+import net.dv8tion.jda.api.components.selections.StringSelectMenu;
 import net.dv8tion.jda.api.interactions.components.selections.StringSelectInteraction;
 import net.dv8tion.jda.api.utils.data.DataArray;
 import net.dv8tion.jda.api.utils.data.DataObject;
